There are no memory leaks
There are no memory leaks left in the code, period. None can be found via
valgrind or mudflap and therefore it is leak free. However some of the code
is very messy and the delete code for the errors thrown could be much cleaner.
I will work on improving that next.